Fixed Search Page
the correct HTML for the search page was saved as test.ejs. i changed the html in the searchresults file with the html in that page with minor changes. i dont know why it wasnt done on the searchresults file to begin with so i saved the original on my device just in case.